+++ lib/enumerize/base.rb
@@ -79,7 +79,7 @@ module Enumerize
         if attr.kind_of? Multiple
           errors.add attr.name unless value.respond_to?(:all?) && value.all? { |v| v.blank? || attr.find_value(v) }
         else
-          errors.add attr.name unless attr.find_value(value)
+          errors.add attr.name, :inclusion unless attr.find_value(value)
         end
       end
     end