+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### enhancements
   * Add `:only` and `except` options to the `Attribute#options` method. (by [@thehappycoder](https://github.com/thehappycoder))
-  * ActiveRecord scopes. (by [@lest](https://github.com/lest))
+  * ActiveRecord scopes. (by [@lest](https://github.com/lest) and [@banyan](https://github.com/banyan))
   * Support for RailsAdmin (by [@drewda](https://github.com/drewda))
 
 ### bug fix